deleted the code of legacy parser, made the symbolic parser the default path, created and tested on two test files to confirm parser's correctness
